リマインド削除処理実装 #2
